fix false positives for duplicate-valued enums (#4)
See issue #3.

Some enums (such as those generated by facebook/ent) have the form:

    // AccessControl defines the type for the access_control enum field.
    type AccessControl string

    // AccessControlAll is the default AccessControl.
    const DefaultAccessControl = AccessControlAll

    // AccessControl values.
    const (
            AccessControlAll       AccessControl = "all"
            AccessControlAny       AccessControl = "any"
            AccessControlAnonymous AccessControl = "anonymous"
    )

When combined with a switch statement of the form:

    switch ac {
    case AccessControlAll, AccessControlAny, AccessControlAnonymous:
    }

this results in false positive report from exhaustive:

    missing cases in switch of type AccessControl: DefaultAccessControl

To address this, this change takes into account the values of the enum
members. If an enum member of a particular value is listed in a switch
statement case, the enum member and any other enum member with the same
value is considered to be accounted for.

This would mean that in the above example the existence of AccessControlAll
in the switch would mean that DefaultAccessControl, which has the same
"all" value, is to also be considered accounted for. So the false
positive report will not be generated by the exhaustive program.

If the switch was of the form:

    switch ac {
    case AccessControlAny:
    }

Then the exhaustive command now reports:

    missing cases in switch of type AccessControl:
    DefaultAccessControl|AccessControlAll, AccessControlAnonymous

The '|' indicates that either DefaultAccessControl or AccessControlAll
is required to be present (they have the same value).

In cases in which the value of a enum member cannot be determine using
go/analysis or Go standard libarary provided info---this can happen for
iota enum members---then only the single listed enum member is considered
accounted for (similar to old behavior).

---

Also: remove maps support entirely, including maps testdata