refactor and unit test switch.go; various other changes (#19)
### Switch.go refactor and testing

Refactor code in switch.go to be unit-testable. Add some unit tests in switch_test.go.

### Deprecate -ignore-pattern flag

The -ignore-pattern flag is deprecated and has no effect. Add -ignore-enum-members,
which has the same behavior, instead. Update the docs accordingly.

### Support checking strategies

The new -checking-strategy flag can be used to specify a checking strategy, one of
"value" or "name". The default strategy is "value", which retains the existing behavior,
of requiring every enum member _value_ to be listed in a switch statement. 
The "name" strategy requires every enum member _name_ to be listed in a switch 
statement. 

Add related docs in doc.go.

Add tests for this change in the testdata packages "duplicateenumvalue/...".

### Analyzer does not suggest fixes

The "-fix" flag is now a no-op. The analyzer no longer reports any suggested fixes.
Also remove source code and testdata related to this flag. Update doc.go
accordingly.

### CI

Include Go 1.14 explicitly in CI matrix. (This is to match with the go version in go.mod.)

### Repo documentation

Add a separate CONTRIBUTING file. General corrections and updates to README.

### Other changes

Improve gob-compatibility testing of fact types.