ability to ignore specific types through new -ignore-enum-types flag
This commit adds an -ignore-enum-types flag to the analyzer. The value
is a regexp pattern that specifies types that should be ignored in
exhaustiveness checks. The new flag is similar to the existing
-ignore-enum-members flag, except that the new flag is used to ignore
types, not constants.

Also: update and clarify related documentation. Re-add the "Skip
analysis" section that was removed in earlier commits.

Related test files are added to testdata/src/ignore-pattern/, which is a
rename of the previously named ignore-enum-member directory.