faster test running times by simplifying imports and using -testing.short
This commit improves uncached test running times for the full suite from
around 5 seconds on my machine to around 3 seconds. Most of the
improvement is due to removing imports to more complex packages, such as
fmt and log, in the testdata files.

With the -short flag: The dedicated complex packages usage test is moved
out of general and into its own package, and placed behind the -short
flag. This makes running tests with -short even faster at around 1.5
seconds on my machine.

The Makefile adds a testshort target.