account for anonymous interface in type params
Previously the typeparams logic did not account for anonymous interfaces
in type paramters. Fix this, and add tests.

A bit of housekeeping to ease debugging: The types and funcs in the
testdata typeparams package are moved to separate files. This should
ease debugging. To debug a problematic test case in isolation, the
single function for the test case can be added to a separate new file
(say funcs_debug.go), and the existing funcs in funcs.go can be disabled
altogether by setting build constraint in funcs.go to "go118 && never",
such that it isn't included.

Also, move the check for same basic kind-ness to once at the end.